refactor(core): extract shared stream-events factory
Lift the shared child-process stream state machine into
providers/shared/stream-events.ts (streamChildAsTurnEvents) and wire
all four adapters (claude/cursor/codex/gemini) through it.

codex+gemini synthetic done event preserved via the onClose hook.
Factory body is byte-identical to the prior per-adapter state machine.
